refactor: make multi select field a reusable component
